refactor(attach file): change for api folder
